refactor(analysis-types): extract technical debt DTO
